+++ libs/adapter/index.ts
@@ -16,7 +16,7 @@ export * from './file-store/url-file-store';
 export * from './logging-service/mongo-db-logging-service';
 export * from './mongo-db/mongo-db-helper';
 
-export * from './nest/graphql/graphql-auth-guard';
+export * from './nest/graphql/graphql-jwt-auth-guard';
 export * from './nest/message/message-controller';
 export * from './nest/message/message-module';