+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.0.68
 
 - feat: add public-key-challenge-store and types
-- feat: add graphql-auth-guard
+- feat: add graphql-jwt-auth-guard
 - feat: add eth-auth-service
 
 ## 0.0.67